+++ backend/src/test/java/com/clothes/perst/PersonalColor/SortTests.java
@@ -1,4 +1,4 @@
-package com.clothes.perst;
+package com.clothes.perst.PersonalColor;
 
 import org.springframework.boot.test.context.SpringBootTest;
